Removed comms.py by importing comms from HoloViews
